requirements.txt
discord.py[voice]
weebapi
simplejson
